Major Code refactor (#116)
Modularise.
Split `extra_files_parser.py` into components 

Fixes #115 
Fixes #106
