+++ castep_outputs/test/test_dumpers.py
@@ -1,13 +1,13 @@
 # pylint: skip-file
 
-import unittest
 import io
 import json
 import pprint
+import unittest
 from pathlib import Path
 
-from castep_outputs.castep_outputs_main import (parse_all, parse_single)
-from castep_outputs.utility import normalise
+from castep_outputs.cli.castep_outputs_main import parse_all, parse_single
+from castep_outputs.utilities.utility import normalise
 
 try:
     from ruamel import yaml
