+++ castep_outputs/utilities/datatypes.py
