+++ pyproject.toml
@@ -33,7 +33,7 @@ yaml = ["pyYAML>=3.13"]
 docs = ["jupyter-book>=0.13.1", "sphinx-book-theme>=0.3.3", "sphinx-argparse>=0.4.0"]
 
 [project.scripts]
-castep_outputs = "castep_outputs.castep_outputs_main:main"
+castep_outputs = "castep_outputs.cli.castep_outputs_main:main"
 
 [project.urls]
 Homepage="https://github.com/oerc0122/castep_outputs"
